+++ test/OpenTelemetry.Instrumentation.Hangfire.Tests/DependencyInjectionConfigTests.cs
@@ -3,6 +3,7 @@
 
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
 using Xunit;