+++ src/OpenTelemetry.Instrumentation.Hangfire/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Add metrics instrumentation following workflow semantic conventions.
+  ([#3258](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3258))
+
 ## 1.13.0-beta.1
 
 Released 2025-Oct-23